Update setup.py
So for everyone wonderning why I spent an hour only updating setup.py, 
if you require a package of which your needed version only exists on 
Github but there is one with the same package name in Pypi, things don't 
work. subprocess seems to be an ok-ish workaround and I'm tired.